+++ CHANGELOG.md
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - added symfony/event-dispatcher ^3.0.0 requirement
 - added optional feature to listen events from RequestSender
 ### Deprecated
-- Method SubscriberInfo::addPropertyChunk is deprecated, use SubscriberInfo::addProperty instead  
+- method SubscriberInfo::addPropertyChunk is deprecated, use SubscriberInfo::addProperty instead
+- method SpecificCsvMethodResponse::getCsvLinesWithoutHeader is deprecated, use SpecificCsvMethodResponse::getCsvReader instead
 
 ## [0.4.0] - 2017-08-09
 Fully refactored library, changed everything